pkgdown/custom.css
pkgdown/home.html

Check the number of crowns included in each image

As the extend of images can changed from one image to another, this function return the number of crowns included in each image and the vector of the id contained for each id.

Usage

check_availableCrowns(path_bbox, crownsFile, dates = NULL)

Arguments

path_bbox

The path to the non NA Bbox return by the function extract_bboxImages()

crownsFile

sf object for the crowns with an 'id' variable.

dates

chr. Vector with dates (format should be '%Y%m%d', p.e '20220925'). The order of the dates should match with the order of the dates of the image in the path_bbox

Value

list and dotchart

pkgdown/footer.html