Extract RGB metrics

The function extracts RGB metrics (red, green, blue, sumrgb, rcc, gcc, bcc, gndvi, gli) for each crown at each date. It first compute the metrics for the whole mosaics and then extracts the value at the crown scale using the exactextractr::exact_extract function. The mean and / or the variance can be extracted (see 'fun' parameter).

Usage

extract_rgbValues(
  path_images,
  path_crowns,
  out_dir_path,
  ncor = 1,
  sites = NULL,
  dates = NULL
)

Arguments

path_images

a list with the full paths to the RGB rasters.

path_crowns

chr. Path to the crown file

out_dir_path

chr. The path to the directory use to stored the result.

ncor

num. Number of cores

sites

chr. name of the site, p.e 'Mbalmayo'.

dates

chr. vector of dates (format should be 'yyyy_mm_dd', p.e '2022_09_25'). The order of the dates should match with the order of the path_images !