pkgdown/custom.css
pkgdown/home.html

Plot the heatmap of Labels

A fct function

Usage

heatmap_Labels(
  longLabels,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL,
  simplify = FALSE,
  repro = FALSE
)

Arguments

longLabels

tibble or dataframe of long data format for labels

Species

chr. Specifying the species you want to filter

Genus

chr. Specifying the genus you want to filter

Family

chr. Specifying the family(ies) you want to filter

title

chr. The title of the plot

simplify

logical. When TRUE, the plot will use simplified labels instead of raw labels.

repro

logical. When TRUE, the flowers and fruits observations will be add to the plot.

Value

return a ggplot

pkgdown/footer.html