pkgdown/custom.css
pkgdown/home.html

Merge labels and rgb values

A function to merge the labels from long format and the rgb metrics values.

Usage

merge_values(
  data_labeling,
  rgb_data,
  out_dir_path = NULL,
  file_type = ".RData"
)

Arguments

data_labeling

tibble or dataframe which contains the labels directly import from the xlsx file.

rgb_data

tibble or dataframe with the rgb metrics values.

Value

A tibble with the variable site, id, date, family, genus, species, phenophase, type, metric, band, value, obs, comments, update, Usable_crown.

Examples

library(canObsR)

data('"data_labeling"')
#> Warning: data set ‘"data_labeling"’ not found
data('data_labeling')
merge_values(data_labeling,rgb_data)
#> # A tibble: 263,536 × 23
#>    site       id date       family   genus species phenophase type  metric band 
#>    <fct>   <int> <date>     <fct>    <fct> <fct>   <chr>      <fct> <fct>  <fct>
#>  1 Bouamir   894 2022-03-26 Combret… Term… Termin… L/D        RGB   mean   red  
#>  2 Bouamir   895 2022-03-26 Combret… Term… Termin… D/F        RGB   mean   red  
#>  3 Bouamir   896 2022-03-26 Fabaceae Hylo… Hylode… NA         RGB   mean   red  
#>  4 Bouamir   897 2022-03-26 Irvingi… Desb… Desbor… NA         RGB   mean   red  
#>  5 Bouamir   898 2022-03-26 Malvace… Dubo… Dubosc… L/D        RGB   mean   red  
#>  6 Bouamir   899 2022-03-26 Fabaceae Albi… Albizi… D          RGB   mean   red  
#>  7 Bouamir   900 2022-03-26 Fabaceae Albi… Albizi… D          RGB   mean   red  
#>  8 Bouamir   901 2022-03-26 Lecythi… Pete… Peters… L          RGB   mean   red  
#>  9 Bouamir   930 2022-03-26 Irvingi… Desb… Desbor… NA         RGB   mean   red  
#> 10 Bouamir   931 2022-03-26 Fabaceae Pter… Pteroc… L          RGB   mean   red  
#> # ℹ 263,526 more rows
#> # ℹ 13 more variables: value <dbl>, PPfoliar1 <chr>, PPfoliar2 <chr>,
#> #   PPFlo <dbl>, PPFr <dbl>, PPFlo_uncertainty <dbl>, PPFr_uncertainty <dbl>,
#> #   desynchr <dbl>, PPfoliar2_uncertainty <dbl>, obs <chr>, comments <chr>,
#> #   update <chr>, Usable_crown <chr>

pkgdown/footer.html