pkgdown/custom.css
pkgdown/home.html

Labeling shiny app

Shiny app to do the labeling based on crowns images.

Shiny app to do the visualise the crowns data (labels and rgb indices).

Usage

shiny_labels(data_labeling)

shiny_labels(data_labeling)

Arguments

data_labeling

tibble. Labeling file resulting from the create_labelingFile

Examples

library(canObsR)

data(data_labeling)
data_labeling
#> # A tibble: 18,900 × 12
#>    site       id family   genus species     n obs   update date       phenophase
#>    <chr>   <int> <chr>    <chr> <chr>   <int> <chr> <chr>  <date>     <chr>     
#>  1 Bouamir   897 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  2 Bouamir   930 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  3 Bouamir   935 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  4 Bouamir   936 Irvingi… Desb… Desbor…   837 NB    2024_… 2022-03-26 L         
#>  5 Bouamir   940 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  6 Bouamir   945 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  7 Bouamir   952 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  8 Bouamir   965 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#>  9 Bouamir   980 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#> 10 Bouamir   987 Irvingi… Desb… Desbor…   837 NA    NA     2022-03-26 NA        
#> # ℹ 18,890 more rows
#> # ℹ 2 more variables: comments <chr>, Usable_crown <chr>
# shiny_labels(data_labeling = data_labeling) # Run it

library(canObsR)

# shiny_visualisation() # Run it

pkgdown/footer.html